Only output single currency breadcrumbs when rendering the settings page (#2284)
* Only output single currency breadcrumbs when rendering the settings page

The 'save' method calls 'get_settings', causing the breadcrumbs to
be rendered also when submitting the single currency settings form.

This commit adds a check to prevent rendering the breadcrumbs when
not rendering the settings page, avoiding the warning and extra
breadcrumbs at the top of the page.

* Move single currency breadcrumbs to the woocommerce_settings_ hook

We should avoid outputting in get_settings, as it's also used to get the
settings in non-rendering contexts. This commit moves the breadcrumbs
logic to the woocommerce_settings_wcpay_multi_currency hook and renders
them when viewing a single currency page.