Format customer currency based on the customer's country (#2312)
* Use country-specific formatting when building currency format from locale-info

* Add get_customer_country method to Utils

We need a way to get the customer country to be able to use it
when picking a country-specific currency formatting and also
when automatically selecting a currency.

This currently doesn't take geolocation into account if there's
a customer object, so this will have to be extended later when
working on the currency selection.

* Use customer country in get_currency_format

Some currencies are formatted differently depending on the
country. This commit adds a country check to pick the right
formatting settings from currency_format.

* Add _admin_ to multi-currency register/enqueue scripts

* Update mini cart when country is updated on legacy checkout

When the country is selected in the legacy checkout page, it could
entail in a change to the currency formatting. In those cases, the
mini-cart widget is not updated automatically, leading to different
formatting being used in the order summary and mini cart.

* Fix get_customer_country tests

* Use user locale instead of country when formatting currencies

* Fix get_user_locale_country mock in tests

* Add return types to Frontend_Currencies methods

* Revert "Update mini cart when country is updated on legacy checkout"

This reverts commit cee4f8cd5062f6f7e4c4ec82b373c78c0e4254ee.