Switch the charge ID and the paymentIntent ID in the copy of a couple screens (#2937)
* Display the charge ID instead of the paymentIntent ID in the order notes on success

When a payment is successfully processed, the order notes now displays the charge ID instead of the paymentIntent ID. The URL it links to continues to be the same, only the anchor's text was updated.

* Replace the charge ID with the paymentIntent ID in the top card of the Transaction detail

In the Transaction detail screen, display the paymentIntent ID instead of the charge ID. This is what Stripe displays there on their own dashboard.

* Adjust test for verifying the order notes for a successful payment

Because we're now displaying the charge ID instead if the paymentIntent ID, the existing test fails. Updated so it looks for the charge ID in the order note instead.

* Fallback to the charge ID if there's no paymentIntent ID in the charge object

According to Stripe's docs, a charge object may not have an associated paymentIntent. It's an optional property. I don't see how not having it would be possible in this context, but adding a fallback just in case.

* Update snapshot for when rendering the loading state 1

The Payment ID now contains the paymentIntent ID instead of the charge ID. Its placeholder is 'pi_xxx' instead of 'ch_xxx'. This commit updates the snapshot to expect this change.

* Add changelog entry for switching the Charge ID and PaymentIntent ID in a couple screens