Display specific error message instead of a generic one in the checkout block when payments fail (#3260)
* Throw an exception when non-upe card processing fails

We're currently adding an error notice and returning a 'failure' response with no information about the error. This works fine in the regular checkout, but in the blocks one, the notices added by this method are cleared and a generic error message is displayed instead.

    This commit replaces the error notice with an exception so the error message shows up both in the regular and block checkout.

* Throw an Exception when UPE's failure limit is reached

We're now throwing an exception instead of adding an error notice when the UPE's failure limit is reached. This ensures the error message shows up both in the regular and in the block checkout.

* Add changelog entry

* Adjust tests to expect 'process_peyment' to throw an Exception on failure

* Adjust e2e tests related to WC Blocks' UPE card failure messages on checkout

We're not displaying generic messages anymore when UPE cards fail for some reason on checkout. This commit adjusts those e2e tests that are expecting the old generic error to show up.

* Adjust typo in the newly added docblock

* Adjust unit tests that were expecting exceptions where no exceptions were thrown