Remove the 'add new payment methods' item from the tasks list (#3134)
* Register a page for adding the extra payment methods

This duplicates the current view for the task 'woocommerce-payments--additional-payment-methods', which is going to be removed after adjusting the links and redirections to the new page.

* Redirect to the new additional-payment-methods page instead of the task one

When enabling UPE via the inbox notice or in the settings page, you get redirected to the task page. Instead, go to the newly added page for adding new payment methods.

* Remove the additional payment methods task

The task in the overview list won't be used anymore, and the task page was moved to a regular page. All the redirects to the task page were also moved to the new regular page.

* Adjust e2e tests for the payment methods task

This task shouldn't be expected to be shown anymore in the e2e tests.

* Adjust js tests to expect the updated 'Additional payments' screen

* Adjust the additional-methods-setup test to expect a page and not a task

* Rename file for the additional-methods-setup js tests

* Remove unneeded wrapper for the UpePreviewMethodSelector component

* Add changelog entry

Co-authored-by: Dat Hoang <htdat@users.noreply.github.com>
Co-authored-by: Harris Wong <harris.wong@automattic.com>