Remove duplicative multi-currency task (#3436)
* Add a new page to move the multi-currency task's contents into

This commit is just registering a new page and its URL to load WC's scripts. It's loading the additional payment methods page for now.

* Load a component with the contents of the multi-currency task container

This populates the new page for handling the multi-currency setup without a task.

* Remove the unused multi-currency task component and variables

Because we're converting this task to a page, there's no need of the task component and other arguments and variables used to set the task up.

* Point the multi-currency note's link to the new page

Now it takes you to the URL using the 'path' argument instead of the 'task' one.

* Adjust js tests that depended on the removed multi-currency task

* Remove the tests related to the removed multi-currency task

The multi-currency setup is now a page, the tests for the task will now fail thus removing it. We can probably remove this set of tests unless we add some others as we aren't really testing much here.

* Rename file for the multi-currency setup tests

* Update the URL of the multi-currency setup page

I had set it to 'payments/multi-currency-available' but keeping it similar to what it was before makes more sense. I also don't recall the reasoning behind the 'available' URL approach.

* Change the title of the multi-currency setup page

Changing it from 'Home' that it was before, to 'Set up multiple currencies' so it's more descriptive.

* Adjust the MultiCurrencySetupPage test so it refers to a page instead of a task

We're moving away from it being a task, thus replacing the references of 'task' with 'page' in the test.

* Add changelog entry

* Add changelog entry to readme.txt