Remove leftovers of the "Boost your sales by accepting new payment methods" task (#3443)
* Remove unused additionalMethodsSetup.isTaskVisible variable

* Display nothing in the additional-payment-methods page when UPE preview is disabled

When UPE preview is disabled, there shouldn't be any link nor redirects taking the user to this page. Because no one would naturally get to this page when UPE preview is disabled, we're displaying nothing in this case.

* Remove unused MethodsSelector component

We don't show anything in the additional-payment-methods page which is where this component was used, thus removing it.

* Display the UPE additional-payment-methods wizard even when the UPE preview is disabled

When the UPE preview is disabled, there shouldn't be a way to get to this page. But there are instances where the inbox note for setting up additional payment methods continues to show up when the preview option is updated in the database, as if it were cached. So I'm opting for displaying the wizard to prevent WCPay from looking broken if this inbox note is displayed in this case. Ideally, the inbox note wouldn't be displayed and there wouldn't be a way to get to the additional-payment-methods setup page. This is an edgy case, though. UPE preview is enabled by default.

* Remove unused isSetupCompleted localized JS variable

This variable was used for when displaying the task in the task list. We've removed it so its value isn't relevant anymore.

* Remove JS tests for the deleted MethodsSelector component

* Remove PHPUnit tests for a removed method

The method WC_Payments_Admin::is_page_eligible_for_additional_methods_setup_task() was removed in previous commits because the additional-methods-setup task was removed, thus removing the PHPUnit tests for this method.

* Remove unused property in the additionalMethodsSetup test

This isn't used in the actual flow, and it's not relevant to the tests either. Thus removing it.