Cache UPE intents for reuse (#3684)
* Cache UPE intents in cookies

* Keep cookie constants at one place

Using the getConfig utility, PHP constants can be accessed in JS.

* Fix payment intent cookie with setup intent value

This happens when checking out a subscription product that has a free trial.

* Remove cookies more effectively

Using unset along with setcookie increases the likelihood of clean
removal of cookies.

This also fixes the issue where cookies weren't deleted on using a new
payment method on the Checkout page.

* Decrease cookie expiry time to 10 mins

This minimizes chances of a stale cookie, that could not be successfully
removed, from causing nasty errors.

* Remove UPE cookies on adding a new payment method

The hook woocommerce_after_account_payment_methods is only used as an
approximation. It is triggered on load of the Payment Methods page in
My Account. This is the page where user is redirected after a new payment
method is successfully added.

* Split remove cookies method into separate methods

* Remove cookie on failure of redirection payment

Giropay, SEPA, Sofort, iDEAL, etc.

* Remove cookie on failure of a saved payment method

* Fix the UPE gateway import

* Remove cookie on failures captured on client-side

Client-side = JS

Examples of such errors include:
* a new card that is declined
* a new card that requires authentication and eventually fails it

Such failures are not captured on server-side (PHP code) because these
scenarios are directly handled by Stripe's JS library.

* Use wc_setcookie for removing UPE cookies

* Make payment intent id prefix more robust

* Update changelog