Add support for platform checkout server side payment processing (#3811)
* Add info necessary to pay via platform checkout

* Send account ID with platform init request

* Handle update_order when payment method is "null"

There seem to be some cases where the payment method id is passed in as
the string "null" instead of the value null. Pretty weird, but easy to
handle.

* Add handling for payment via platform checkout

* Add filter to change account ID for intent confirmation

This limits the use of the filter to only very specific situations
necessary for platform checkout to work. The filter is useless if the
account ID doesn't match the account ID used to create the payment
intent.

* add changelog entries

* Clean up intent confirmation JS

* Sanitize platform checkout intent parameter

We use sanitize_user since it seems the most appropriate given which
characters Stripe IDs may contain.

Discussion on this can be found in
https://github.com/Automattic/woocommerce-payments/pull/3811#discussion_r808674621