Fix KYC treatment mode API and menu issues (#4174)
* Avoid redirecting settings to onboarding

We added it with the intention to redirect WC task paths directly to the onboarding page without going to connect. But we found now that it causes issues with the settings menu being redirected to onboard even after being onboarded.

* Redirect directly to onboarding page if come from WC Admin task and are in treatment mode

* Use user token for new onboarding requests

* Add changelog entry

* Add missing API Client tests

* Ensure the server is connected before redirecting

To be able to make API requests in the KYC form we need to be connected to jetpack. Check it before redirecting to the onboarding page, to start the server connection if not.

* Update experiment id