Fix customer not found errors after site migration (#5383)
* Fix customer not found errors after site migration

After a site has been migrated to a different WCPay account pre-existing
account will be associated with Stripe customer objects that don't
exist. This will result in a "No such customer" error in UPE checkouts
and a "We're not able to process this request" error in classic
checkouts.

To fix the issue we make the customer update function synchronous
instead of asynchronous. This rolls back a significant performance
update, but WCPay stores shouldn't hit this path most of the time
anyway, so it shouldn't matter for most stores.