Update UPE logic around the cards payment method and improve onboarding styling (#5296)
* Add required and locked support for payment methods list component

The `required` prop displays a  `(Required)` secondary label, and
the `locked` prop will disable the checkbox input and use an
appropriate cursor.
The combination of these two props allows us to have payment
methods that are active by default and can't be deactivated.

* Add required and locked support for payment methods checkboxes component

The `required` prop displays a  `Required` secondary label, and
the `locked` prop will disable the checkbox input and use an
appropriate cursor.
The combination of these two props allows us to have payment
methods that are active by default and can't be deactivated.

* Require and lock the card payment method in settings for UPE

When UPE is enable, the card payment method needs to be active and
it can't be deactivated. We inform the user that the card is
required and disable the checkbox control to prevent the user from
deactivating it.

* Add pre-selected card payment method for UPE wizard

The card payment method is mandatory when activating UPE and
it should be shown to the user as such (checked and disabled since
it can't be deactivated).

* Adjust styling of the UPE activation wizard to better match original designs

See https://github.com/Automattic/woocommerce-payments/issues/4114#issuecomment-1344101862
for the target designs.

Mostly the spacing was off in certain places and the payment method title text was
a little bit large.

* test: Add tests for payment methods list  Required and Locked behaviors

* test: Fix tests for payment methods checkboxes

* test: Add tests for payment methods checkbox Required and Locked behaviors

* test: Fix tests for UPE add payment methods wizard

* Add changelog for #4114

* Better naming for payment method constant flag

Since we are in a component context, there is no need to repeat the component name.

* Reduce styling changes to payment methods

This way we reduce future refactoring work and avoid further making the current situation worse.